requirements.txt
qwen-tts
flask>=3.0.0
flask-cors>=4.0.0
soundfile>=0.12.0
torch>=2.0.0
openai-whisper
numpy
pydub
